+++ src/tko/play/tasktree.py
@@ -620,7 +620,7 @@ class TaskTree:
         lang = self.rep.get_lang()
         if folder is None:
             return False
-        drafts = Drafts.load_drafs(folder, lang)
+        drafts = Drafts.load_md_and_drafs(folder, lang)
         if drafts:
             return True
         return False